requirements.txt
yahoo-fantasy-api>=2.1.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests-html>=0.10.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
